+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/docker_kurtosis_backend_enclave_functions.go
@@ -44,10 +44,6 @@ const (
 
 
 )
-var (
-	enclaveCreationTimeRetroCompatibilityCheckDeadline = time.Date(2023, 1,1,0,0,0,0,time.Local)
-)
-
 
 // TODO: MIGRATE THIS FOLDER TO USE STRUCTURE OF USER_SERVICE_FUNCTIONS MODULE
 
@@ -897,12 +893,10 @@ func getEnclaveCreationTimeFromNetwork(network *types.Network) (*time.Time, erro
 	labels := network.GetLabels()
 	enclaveCreationTimeStr, found := labels[label_key_consts.EnclaveCreationTimeLabelKey.GetString()]
 	if !found {
-		//TODO remove this condition after 2023-01-01 when we are sure that there is not any old enclave created without the creation time label
 		//Handling retro-compatibility, enclaves that did not track enclave's creation time
-		if time.Now().Before(enclaveCreationTimeRetroCompatibilityCheckDeadline){
-			return nil, nil
-		}
-		return nil, stacktrace.NewError("Expected to find network's label with key '%v' but none was found", label_key_consts.EnclaveCreationTimeLabelKey.GetString())
+		return nil, nil //TODO remove this return after 2023-01-01
+		//TODO uncomment this after 2023-01-01 when we are sure that there is not any old enclave created with the creation time annotation
+		//return nil, stacktrace.NewError("Expected to find network's label with key '%v' but none was found", label_key_consts.EnclaveCreationTimeLabelKey.GetString())
 	}
 
 	enclaveCreationTime, err := time.Parse(time.RFC3339, enclaveCreationTimeStr)